Rename context.rs -> tx.rs
