+++ crates/celo-revm/src/lib.rs
@@ -6,10 +6,10 @@
 extern crate alloc as std;
 
 pub mod chain_info;
-pub mod context;
 pub mod evm;
 pub mod precompiles;
 pub mod transfer;
+pub mod tx;
 
-pub use context::CeloTxEnv;
 pub use evm::CeloEvm;
+pub use tx::CeloTxEnv;
